requirements.txt
fastapi
jinja2
uvicorn[standard]
arel
python-multipart
sse-starlette
aiofiles
openai
htmlmin